Move upload/download functionality out of cookiecutter, into molo.core
