+++ molo/core/cookiecutter/scaffold/{{cookiecutter.directory}}/{{cookiecutter.app_name}}/urls.py
@@ -9,7 +9,7 @@ from wagtail.wagtailadmin import urls as wagtailadmin_urls
 from wagtail.wagtaildocs import urls as wagtaildocs_urls
 from wagtail.wagtailcore import urls as wagtail_urls
 
-from {{cookiecutter.app_name}} import views
+from molo.core.views import upload_file, download_file
 
 # implement CAS URLs in a production setting
 if settings.ENABLE_SSO:
@@ -24,9 +24,9 @@ else:
 
 urlpatterns += patterns(
     '',
-    url(r'^django-admin/upload_media/', views.upload_file,
+    url(r'^django-admin/upload_media/', upload_file,
         name='molo_upload_media'),
-    url(r'^django-admin/download_media/', views.download_file,
+    url(r'^django-admin/download_media/', download_file,
         name='molo_download_media'),
     url(r'^django-admin/', include(admin.site.urls)),
     url(r'^admin/', include(wagtailadmin_urls)),
