+++ molo/core/forms.py
@@ -1,4 +1,5 @@
 from django.utils import timezone
+from django import forms
 
 from wagtail.wagtailadmin.forms import WagtailAdminPageForm
 
@@ -42,3 +43,17 @@ class ArticlePageForm(WagtailAdminPageForm):
                     )
 
         return cleaned_data
+
+
+class MediaForm(forms.Form):
+    '''Form to upload a sinlge zip file.'''
+    zip_file = forms.FileField(label="Zipped Media File")
+
+    def clean_zip_file(self):
+        file = self.cleaned_data['zip_file']
+
+        if file:
+            extension = file.name.split('.')[-1]
+            if extension != 'zip':
+                raise forms.ValidationError('File Type Is Not .zip')
+        return file
