+++ molo/core/views.py
@@ -1,12 +1,15 @@
-from os import environ
+from os import environ, path, walk
 import pkg_resources
 import requests
+import zipfile
+import StringIO
 
 from django.conf import settings
 from django.contrib import messages
+from django.contrib.auth.decorators import user_passes_test
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
 from django.core.urlresolvers import reverse
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponse, HttpResponseRedirect
 from django.shortcuts import redirect, get_object_or_404, render
 from django.utils.translation import (
     LANGUAGE_SESSION_KEY,
@@ -16,9 +19,10 @@ from django.utils.translation import ugettext as _
 from wagtail.wagtailcore.models import Page
 from wagtail.wagtailsearch.models import Query
 
-from molo.core.utils import generate_slug, get_locale_code
+from molo.core.utils import generate_slug, get_locale_code, replace_media_file
 from molo.core.models import PageTranslation, SiteLanguage, ArticlePage
 from molo.core.known_plugins import known_plugins
+from molo.core.forms import MediaForm
 from django.views.generic import ListView
 
 
@@ -164,3 +168,40 @@ class TagsListView(ListView):
     def get_queryset(self, **kwargs):
         tag = self.kwargs["tag_name"]
         return ArticlePage.objects.filter(tags__name__in=[tag])
+
+
+@user_passes_test(lambda u: u.is_superuser)
+def upload_file(request):
+    '''Upload a Zip File Containing a single file containing media.'''
+    if request.method == 'POST':
+        form = MediaForm(request.POST, request.FILES)
+        if form.is_valid():
+            replace_media_file(request.FILES['zip_file'])
+            return HttpResponseRedirect('/django-admin/')
+    else:
+        form = MediaForm()
+    return render(request, 'admin/upload_media.html', {'form': form})
+
+
+@user_passes_test(lambda u: u.is_superuser)
+def download_file(request):
+    '''Create and download a zip file containing the media file.'''
+    if request.method == "GET":
+        zipfile_name = 'media_%s.zip' % settings.SITE_NAME
+        in_memory_file = StringIO.StringIO()
+
+        media_zipfile = zipfile.ZipFile(in_memory_file, 'w',
+                                        zipfile.ZIP_DEFLATED)
+
+        directory_name = path.split(settings.MEDIA_ROOT)[-1]
+        for root, dirs, files in walk(directory_name):
+            for file in files:
+                media_zipfile.write(path.join(root, file))
+
+        media_zipfile.close()
+
+        resp = HttpResponse(in_memory_file.getvalue(),
+                            content_type="application/x-zip-compressed")
+        resp['Content-Disposition'] = 'attachment; filename=%s' % zipfile_name
+
+        return resp
